+++ changelog.md
@@ -3,7 +3,7 @@
 **What's New**
 * [ErrorHandler]: added to 'promo' skin
 * [App]: page 'Downloads' updated, added illustrations and patterns
-* added useForceUpdate, useUUIError hooks to '@epam/uui' package
+* added useForceUpdate, useUuiError hooks to '@epam/uui' package
 
 **What’s Fixed**
 * [CodeSandbox]: fixed functionality of modal windows and notifications in codesandbox examples