+++ changelog.md
@@ -1,3 +1,13 @@
+# 4.3.1 - xx.xx.xxxx
+
+**What's New**
+* [ErrorHandler]: added to 'promo' skin
+* [App]: page 'Downloads' updated, added illustrations and patterns
+* added useForceUpdate, useUuiError hooks to '@epam/uui' package
+
+**What’s Fixed**
+* [CodeSandbox]: fixed functionality of modal windows and notifications in codesandbox examples
+
 # 4.3.0 - 25.10.2021
 
 **What's New**